ollama_client.py
import requests
import os
from dotenv import load_dotenv

load_dotenv()

OLLAMA_URL = os.getenv("OLLAMA_URL")
MODEL_NAME = os.getenv("MODEL_NAME")
def ask_ollama(context, question):
    prompt = f"""
    Answer the question using ONLY the context below.
    
    Context:
    {context}
    
    Question:
    {question}
    """

    response = requests.post(
        f"{OLLAMA_URL}/api/generate",
        json={
            "model": MODEL_NAME,
            "prompt": prompt,
            "stream": False
        }
    )

    return response.json()["response"]